Replace mixnet backend with mixnet addon
