+++ nodes/nomos-node/Cargo.toml
@@ -20,19 +20,24 @@ overwatch-derive = { git = "https://github.com/logos-co/Overwatch", rev = "ac28d
 tracing = "0.1"
 multiaddr = "0.18"
 nomos-core = { path = "../../nomos-core" }
-nomos-network = { path = "../../nomos-services/network" }
+nomos-network = { path = "../../nomos-services/network", features = ["libp2p"] }
 nomos-api = { path = "../../nomos-services/api" }
 nomos-log = { path = "../../nomos-services/log" }
 nomos-mempool = { path = "../../nomos-services/mempool", features = [
   "mock",
+  "libp2p",
   "metrics",
 ] }
 nomos-metrics = { path = "../../nomos-metrics" }
 nomos-http = { path = "../../nomos-services/http", features = ["http"] }
-carnot-consensus = { path = "../../nomos-services/carnot-consensus" }
+carnot-consensus = { path = "../../nomos-services/carnot-consensus", features = [
+  "libp2p",
+] }
 nomos-storage = { path = "../../nomos-services/storage", features = ["sled"] }
 nomos-libp2p = { path = "../../nomos-libp2p" }
-nomos-da = { path = "../../nomos-services/data-availability" }
+nomos-da = { path = "../../nomos-services/data-availability", features = [
+  "libp2p",
+] }
 nomos-system-sig = { path = "../../nomos-services/system-sig" }
 metrics = { path = "../../nomos-services/metrics", optional = true }
 tracing-subscriber = "0.3"
@@ -53,15 +58,4 @@ tower-http = { version = "0.4", features = ["cors", "trace"] }
 
 [features]
 default = []
-libp2p = [
-  "nomos-network/libp2p",
-  "nomos-mempool/libp2p",
-  "nomos-da/libp2p",
-  "carnot-consensus/libp2p",
-]
-mixnet = [
-  "nomos-network/mixnet",
-  "nomos-mempool/mixnet",
-  "nomos-da/mixnet",
-  "carnot-consensus/mixnet",
-]
+mixnet = ["nomos-network/mixnet"]
