+++ nomos-cli/Cargo.toml
@@ -18,8 +18,10 @@ futures = "0.3"
 tokio = { version = "1", features = ["sync"] }
 overwatch-rs = { git = "https://github.com/logos-co/Overwatch", rev = "2f70806" }
 overwatch-derive = { git = "https://github.com/logos-co/Overwatch", rev = "ac28d01" }
-nomos-network = { path = "../nomos-services/network" }
-nomos-da = { path = "../nomos-services/data-availability" }
+nomos-network = { path = "../nomos-services/network", features = ["libp2p"] }
+nomos-da = { path = "../nomos-services/data-availability", features = [
+  "libp2p",
+] }
 carnot-consensus = { path = "../nomos-services/carnot-consensus" }
 nomos-log = { path = "../nomos-services/log" }
 nomos-libp2p = { path = "../nomos-libp2p" }
@@ -37,8 +39,3 @@ ratatui = "0.24"
 tui-input = "0.8"
 ansi-to-tui = "3"
 rand = "0.8"
-
-[features]
-default = []
-libp2p = ["nomos-network/libp2p", "nomos-da/libp2p"]
-mixnet = ["nomos-network/mixnet", "nomos-da/mixnet"]
