+++ nomos-cli/src/cmds/disseminate/mod.rs
@@ -3,10 +3,7 @@ use crate::da::disseminate::{
 };
 use clap::Args;
 use nomos_log::{LoggerBackend, LoggerSettings};
-#[cfg(feature = "libp2p")]
 use nomos_network::backends::libp2p::Libp2p as NetworkBackend;
-#[cfg(feature = "mixnet")]
-use nomos_network::backends::mixnet::MixnetNetworkBackend as NetworkBackend;
 use nomos_network::NetworkService;
 use overwatch_rs::{overwatch::OverwatchRunner, services::ServiceData};
 use reqwest::Url;
