+++ nomos-services/api/Cargo.toml
@@ -6,8 +6,6 @@ edition = "2021"
 [features]
 default = ["axum"]
 axum = ["dep:axum", "dep:hyper", "dep:tower-http", "utoipa-swagger-ui/axum"]
-libp2p = ["nomos-mempool/libp2p"]
-mixnet = ["nomos-mempool/mixnet"]
 
 [dependencies]
 async-trait = "0.1"
@@ -22,6 +20,7 @@ nomos-network = { path = "../../nomos-services/network" }
 nomos-da = { path = "../../nomos-services/data-availability" }
 nomos-mempool = { path = "../../nomos-services/mempool", features = [
   "mock",
+  "libp2p",
   "openapi",
 ] }
 nomos-metrics = { path = "../../nomos-metrics" }
