+++ nomos-services/carnot-consensus/src/network/adapters/mod.rs
@@ -1,4 +1,4 @@
+#[cfg(feature = "libp2p")]
+pub mod libp2p;
 #[cfg(feature = "mock")]
 pub mod mock;
-#[cfg(any(feature = "libp2p", feature = "mixnet"))]
-pub mod p2p;
