+++ nomos-services/data-availability/src/network/adapters/mod.rs
@@ -1,2 +1,2 @@
-#[cfg(any(feature = "libp2p", feature = "mixnet"))]
-pub mod p2p;
+#[cfg(feature = "libp2p")]
+pub mod libp2p;
