+++ nomos-services/mempool/src/network/adapters/mod.rs
@@ -1,5 +1,5 @@
-#[cfg(any(feature = "libp2p", feature = "mixnet"))]
-pub mod p2p;
+#[cfg(feature = "libp2p")]
+pub mod libp2p;
 
 #[cfg(feature = "mock")]
 pub mod mock;
