+++ nomos-services/network/Cargo.toml
@@ -33,6 +33,6 @@ tokio = { version = "1", features = ["full"] }
 [features]
 default = []
 libp2p = ["nomos-libp2p", "rand", "humantime-serde"]
-mixnet = ["dep:mixnet", "nomos-libp2p", "rand", "humantime-serde"]
+mixnet = ["dep:mixnet"]
 mock = ["rand", "chrono"]
 openapi = ["dep:utoipa", "serde_json"]
