+++ nomos-services/network/src/backends/mod.rs
@@ -2,11 +2,8 @@ use super::*;
 use overwatch_rs::{overwatch::handle::OverwatchHandle, services::state::ServiceState};
 use tokio::sync::broadcast::Receiver;
 
-#[cfg(any(feature = "libp2p", feature = "mixnet"))]
+#[cfg(feature = "libp2p")]
 pub mod libp2p;
-
-#[cfg(feature = "mixnet")]
-pub mod mixnet;
 
 #[cfg(feature = "mock")]
 pub mod mock;
